Added tests for playlists and most rated tracks in MediaRepositoryTest
Now that all tests are finalized and green, some refactoring occurred.
MediaUsageEvent.trackId has been made an INTEGER field.
Unused APIs have been removed.

+++ media/src/main/java/fr/nihilus/music/media/usage/MediaUsageManager.kt
@@ -68,7 +68,7 @@ internal class RxBridgeUsageManager
 
     override fun reportCompletion(trackId: String) {
         scope.launch(Dispatchers.IO) {
-            val singleEventList = listOf(MediaUsageEvent(trackId))
+            val singleEventList = listOf(MediaUsageEvent(trackId.toLong()))
             usageDao.recordUsageEvents(singleEventList)
         }
     }
